déplace le code lié aux imports dans le package `ops`
