+++ src/core/keycloak.go
@@ -10,19 +10,10 @@ import (
 	gocloak "github.com/Nerzal/gocloak/v10"
 	"github.com/gin-gonic/gin"
 	"github.com/golang-jwt/jwt/v4"
-	"github.com/jackc/pgx/v4"
 	"github.com/spf13/viper"
 )
 
-type scope []string
-
-type keycloakUser struct {
-	userName  *string
-	email     *string
-	firstName *string
-	lastName  *string
-	roles     scope
-}
+type Scope []string
 
 func connectKC() gocloak.GoCloak {
 	keycloak := gocloak.NewClient(viper.GetString("keycloakHostname"))
@@ -109,7 +100,7 @@ func fakeCloakMiddleware(c *gin.Context) {
 	c.Next()
 }
 
-func scopeFromClaims(claims *jwt.MapClaims) scope {
+func scopeFromClaims(claims *jwt.MapClaims) Scope {
 	var resourceAccess = make(map[string]interface{})
 	var client = make(map[string]interface{})
 	var scope []interface{}
@@ -137,7 +128,7 @@ func scopeFromClaims(claims *jwt.MapClaims) scope {
 	return tags
 }
 
-func scopeFromContext(c *gin.Context) scope {
+func scopeFromContext(c *gin.Context) Scope {
 	claims, ok := c.Get("claims")
 	if !ok {
 		return nil
@@ -146,7 +137,7 @@ func scopeFromContext(c *gin.Context) scope {
 	return roles
 }
 
-func (sc scope) zoneGeo() []string {
+func (sc Scope) zoneGeo() []string {
 	var zone []string
 	for _, role := range sc {
 		departements := db.GetDepartementForRole(role)
@@ -156,144 +147,6 @@ func (sc scope) zoneGeo() []string {
 		zone = append(zone, s)
 	}
 	return zone
-}
-
-func getKeycloakUsers(c *gin.Context) {
-	userMap, roleMap, err := fetchUsersAndRoles()
-	if err != nil {
-		c.JSON(err.Code(), err.Error())
-		return
-	}
-	err = importUsersAndRoles(userMap, roleMap)
-	if err != nil {
-		log.Print(err.Error())
-		c.JSON(err.Code(), err.Error())
-		return
-	}
-	c.JSON(200, "utilisateurs mis à jour")
-}
-
-func fetchUsersAndRoles() (map[string]keycloakUser, map[string]*string, Jerror) {
-	realm := viper.GetString("keycloakRealm")
-	client := viper.GetString("keycloakClient")
-	kcAdmin := gocloak.NewClient(viper.GetString("keycloakHostname"))
-	jwt, err := kcAdmin.LoginAdmin(
-		context.TODO(),
-		viper.GetString("keycloakAdminUsername"),
-		viper.GetString("keycloakAdminPassword"),
-		realm,
-	)
-	if err != nil {
-		return nil, nil, errorToJSON(500, err)
-	}
-
-	var i = 100000
-	users, err := kcAdmin.GetUsers(
-		context.TODO(),
-		jwt.AccessToken,
-		viper.GetString("keycloakRealm"),
-		gocloak.GetUsersParams{
-			Max: &i,
-		},
-	)
-	if err != nil {
-		return nil, nil, errorToJSON(500, err)
-	}
-
-	clients, err := kcAdmin.GetClients(
-		context.Background(),
-		jwt.AccessToken,
-		realm,
-		gocloak.GetClientsParams{
-			ClientID: &client,
-		},
-	)
-	if err != nil {
-		return nil, nil, errorToJSON(500, err)
-	}
-	if len(clients) != 1 {
-		return nil, nil, newJSONerror(500, "client not found")
-	}
-
-	userMap := make(map[string]keycloakUser)
-	for _, u := range users {
-		roles, err := kcAdmin.GetClientRolesByUserID(
-			context.TODO(),
-			jwt.AccessToken,
-			realm,
-			*clients[0].ID,
-			*u.ID,
-		)
-		if err != nil {
-			return nil, nil, errorToJSON(500, err)
-		}
-		var userRoles []string
-		for _, r := range roles {
-			userRoles = append(userRoles, *r.Name)
-		}
-		userMap[*u.ID] = keycloakUser{
-			userName:  u.Username,
-			email:     u.Email,
-			firstName: u.FirstName,
-			lastName:  u.LastName,
-			roles:     userRoles,
-		}
-	}
-
-	if len(userMap) == 0 {
-		return nil, nil, newJSONerror(204, "no users")
-	}
-
-	roles, err := kcAdmin.GetClientRoles(
-		context.Background(),
-		jwt.AccessToken,
-		realm,
-		*clients[0].ID,
-		gocloak.GetRoleParams{})
-
-	if err != nil {
-		return nil, nil, errorToJSON(500, err)
-	}
-
-	roleMap := make(map[string]*string)
-	for _, r := range roles {
-		roleMap[*r.Name] = r.Description
-	}
-
-	return userMap, roleMap, nil
-}
-
-func importUsersAndRoles(userMap map[string]keycloakUser, roleMap map[string]*string) Jerror {
-	tx, err := db.Get().Begin(context.Background())
-	if err != nil {
-		return errorToJSON(500, err)
-	}
-	defer tx.Commit(context.Background())
-
-	batch := pgx.Batch{}
-
-	batch.Queue("delete from users;")
-	batch.Queue("delete from roles;")
-
-	sqlUser := `insert into users (id, userName, firstName, lastName, roles)
-		values ($1, $2, $3, $4, $5);`
-	for id, user := range userMap {
-		batch.Queue(sqlUser, id, user.userName, user.firstName, user.lastName, user.roles)
-	}
-
-	sqlRole := `insert into roles (code, description)
-		values ($1, $2);`
-	for code, description := range roleMap {
-		batch.Queue(sqlRole, code, description)
-	}
-
-	b := tx.SendBatch(context.Background(), &batch)
-	err = b.Close()
-	if err != nil {
-		return errorToJSON(500, err)
-	}
-
-	return nil
 }
 
 type user struct {
