+++ src/ops/import.go
@@ -1,10 +1,11 @@
-package core
+package ops
 
 import (
 	"compress/gzip"
 	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/signaux-faibles/datapi/src/core"
 	"github.com/signaux-faibles/datapi/src/db"
 	"io"
 	"log"
@@ -100,13 +101,13 @@ type delai struct {
 type entreprise struct {
 	ID    string `json:"_id"`
 	Value struct {
-		Sirets     []string    `json:"sirets" hash:"-"`
-		Siren      string      `json:"siren"`
-		Diane      []diane     `json:"diane"`
-		BDF        []bdf       `json:"bdf"`
-		Ellisphere *ellisphere `json:"ellisphere"`
-		SireneUL   sireneUL    `json:"sirene_ul"`
-		Paydex     *[]paydex   `json:"paydex"`
+		Sirets     []string         `json:"sirets" hash:"-"`
+		Siren      string           `json:"siren"`
+		Diane      []core.Diane     `json:"diane"`
+		BDF        []core.Bdf       `json:"bdf"`
+		Ellisphere *core.Ellisphere `json:"ellisphere"`
+		SireneUL   sireneUL         `json:"sirene_ul"`
+		Paydex     *[]paydex        `json:"paydex"`
 	} `bson:"value"`
 }
 
@@ -130,21 +131,6 @@ type sireneUL struct {
 	NomUsageUniteLegale string     `json:"nom_usage_unite_legale,omitempty"`
 	StatutJuridique     string     `json:"statut_juridique"`
 	Creation            *time.Time `json:"date_creation,omitempty"`
-}
-
-type ellisphere struct {
-	Siren               string  `json:"-"`
-	CodeGroupe          string  `json:"code_groupe,omitempty"`
-	SirenGroupe         string  `json:"siren_groupe,omitempty"`
-	RefIDGroupe         string  `json:"refid_groupe,omitempty"`
-	RaisocGroupe        string  `json:"raison_sociale_groupe,omitempty"`
-	AdresseGroupe       string  `json:"adresse_groupe,omitempty"`
-	PersonnePouMGroupe  string  `json:"personne_pou_m_groupe,omitempty"`
-	NiveauDetention     int     `json:"niveau_detention,omitempty"`
-	PartFinanciere      float64 `json:"part_financiere,omitempty"`
-	CodeFiliere         string  `json:"code_filiere,omitempty"`
-	RefIDFiliere        string  `json:"refid_filiere,omitempty"`
-	PersonnePouMFiliere string  `json:"personne_pou_m_filiere,omitempty"`
 }
 
 // APConso detail
@@ -175,101 +161,6 @@ type apDemande struct {
 	MotifRecoursSE     int     `json:"motif_recours_se"`
 	HeureConsomme      float64 `json:"heure_consommee"`
 	MontantConsomme    float64 `json:"montant_consommee"`
-}
-
-type bdf struct {
-	Siren               string    `json:"siren,omitempty"`
-	Annee               int       `json:"annee_bdf"`
-	ArreteBilan         time.Time `json:"arrete_bilan_bdf"`
-	DelaiFournisseur    float64   `json:"delai_fournisseur"`
-	DetteFiscale        float64   `json:"dette_fiscale"`
-	FinancierCourtTerme float64   `json:"financier_court_terme"`
-	FraisFinancier      float64   `json:"FraisFinancier"`
-	PoidsFrng           float64   `json:"poids_frng"`
-	TauxMarge           float64   `json:"taux_marge"`
-}
-
-type diane struct {
-	NumeroSiren                     string    `json:"numero_siren,omitempty"`
-	ArreteBilan                     time.Time `json:"arrete_bilan_diane,omitempty"`
-	AchatMarchandises               *float64  `json:"achat_marchandises,omitempty"`
-	AchatMatieresPremieres          *float64  `json:"achat_matieres_premieres,omitempty"`
-	AutonomieFinanciere             *float64  `json:"autonomie_financiere,omitempty"`
-	AutresAchatsChargesExternes     *float64  `json:"autres_achats_charges_externes,omitempty"`
-	AutresProduitsChargesReprises   *float64  `json:"autres_produits_charges_reprises,omitempty"`
-	BeneficeOuPerte                 *float64  `json:"benefice_ou_perte,omitempty"`
-	CAExportation                   *float64  `json:"ca_exportation,omitempty"`
-	CapaciteAutofinancement         *float64  `json:"capacite_autofinancement,omitempty"`
-	CapaciteRemboursement           *float64  `json:"capacite_remboursement,omitempty"`
-	CAparEffectif                   *float64  `json:"ca_par_effectif,omitempty"`
-	ChargeExceptionnelle            *float64  `json:"charge_exceptionnelle,omitempty"`
-	ChargePersonnel                 *float64  `json:"charge_personnel,omitempty"`
-	ChargesFinancieres              *float64  `json:"charges_financieres,omitempty"`
-	ChiffreAffaire                  *float64  `json:"ca,omitempty"`
-	ConcesBrevEtDroitsSim           *float64  `json:"conces_brev_et_droits_sim,omitempty"`
-	ConcoursBancaireCourant         *float64  `json:"concours_bancaire_courant,omitempty"`
-	Consommation                    *float64  `json:"consommation,omitempty"`
-	CouvertureCaBesoinFdr           *float64  `json:"couverture_ca_besoin_fdr,omitempty"`
-	CouvertureCaFdr                 *float64  `json:"couverture_ca_fdr,omitempty"`
-	CreditClient                    *float64  `json:"credit_client,omitempty"`
-	CreditFournisseur               *float64  `json:"credit_fournisseur,omitempty"`
-	DegreImmoCorporelle             *float64  `json:"degre_immo_corporelle,omitempty"`
-	DetteFiscaleEtSociale           *float64  `json:"dette_fiscale_et_sociale,omitempty"`
-	DotationAmortissement           *float64  `json:"dotation_amortissement,omitempty"`
-	EffectifConsolide               *int      `json:"effectif_consolide,omitempty"`
-	EfficaciteEconomique            *float64  `json:"efficacite_economique,omitempty"`
-	Endettement                     *float64  `json:"endettement,omitempty"`
-	EndettementGlobal               *float64  `json:"endettement_global,omitempty"`
-	EquilibreFinancier              *float64  `json:"equilibre_financier,omitempty"`
-	ExcedentBrutDExploitation       *float64  `json:"excedent_brut_d_exploitation,omitempty"`
-	Exercice                        float64   `json:"exercice_diane,omitempty"`
-	Exportation                     *float64  `json:"exportation,omitempty"`
-	FinancementActifCirculant       *float64  `json:"financement_actif_circulant,omitempty"`
-	FraisDeRetD                     *float64  `json:"frais_de_RetD,omitempty"`
-	ImpotBenefice                   *float64  `json:"impot_benefice,omitempty"`
-	ImpotsTaxes                     *float64  `json:"impots_taxes,omitempty"`
-	IndependanceFinanciere          *float64  `json:"independance_financiere,omitempty"`
-	Interets                        *float64  `json:"interets,omitempty"`
-	LiquiditeGenerale               *float64  `json:"liquidite_generale,omitempty"`
-	LiquiditeReduite                *float64  `json:"liquidite_reduite,omitempty"`
-	MargeCommerciale                *float64  `json:"marge_commerciale,omitempty"`
-	NombreEtabSecondaire            *int      `json:"nombre_etab_secondaire,omitempty"`
-	NombreFiliale                   *int      `json:"nombre_filiale,omitempty"`
-	NombreMois                      *int      `json:"nombre_mois,omitempty"`
-	OperationsCommun                *float64  `json:"operations_commun,omitempty"`
-	PartAutofinancement             *float64  `json:"part_autofinancement,omitempty"`
-	PartEtat                        *float64  `json:"part_etat,omitempty"`
-	PartPreteur                     *float64  `json:"part_preteur,omitempty"`
-	PartSalaries                    *float64  `json:"part_salaries,omitempty"`
-	ParticipationSalaries           *float64  `json:"participation_salaries,omitempty"`
-	Performance                     *float64  `json:"performance,omitempty"`
-	PoidsBFRExploitation            *float64  `json:"poids_bfr_exploitation,omitempty"`
-	ProcedureCollective             bool      `json:"procedure_collective,omitempty"`
-	Production                      *float64  `json:"production,omitempty"`
-	ProductiviteCapitalFinancier    *float64  `json:"productivite_capital_financier,omitempty"`
-	ProductiviteCapitalInvesti      *float64  `json:"productivite_capital_investi,omitempty"`
-	ProductivitePotentielProduction *float64  `json:"productivite_potentiel_production,omitempty"`
-	ProduitExceptionnel             *float64  `json:"produit_exceptionnel,omitempty"`
-	ProduitsFinanciers              *float64  `json:"produits_financiers,omitempty"`
-	RendementBrutFondsPropres       *float64  `json:"rendement_brut_fonds_propres,omitempty"`
-	RendementCapitauxPropres        *float64  `json:"rendement_capitaux_propres,omitempty"`
-	RendementRessourcesDurables     *float64  `json:"rendement_ressources_durables,omitempty"`
-	RentabiliteEconomique           *float64  `json:"rentabilite_economique,omitempty"`
-	RentabiliteNette                *float64  `json:"rentabilite_nette,omitempty"`
-	ResultatAvantImpot              *float64  `json:"resultat_avant_impot,omitempty"`
-	ResultatExploitation            *float64  `json:"resultat_expl"`
-	RotationStocks                  *float64  `json:"rotation_stocks,omitempty"`
-	StatutJuridique                 string    `json:"statut_juridique,omitempty"`
-	SubventionsDExploitation        *float64  `json:"subventions_d_exploitation,omitempty"`
-	TailleCompoGroupe               *int      `json:"taille_compo_groupe,omitempty"`
-	TauxDInvestissementProductif    *float64  `json:"taux_d_investissement_productif,omitempty"`
-	TauxEndettement                 *float64  `json:"taux_endettement,omitempty"`
-	TauxInteretFinancier            *float64  `json:"taux_interet_financier,omitempty"`
-	TauxInteretSurCA                *float64  `json:"taux_interet_sur_ca,omitempty"`
-	TauxMargeCommerciale            *float64  `json:"taux_marge_commerciale,omitempty"`
-	TauxValeurAjoutee               *float64  `json:"taux_valeur_ajoutee,omitempty"`
-	ValeurAjoutee                   *float64  `json:"valeur_ajoutee,omitempty"`
-	NomEntreprise                   string    `json:"nom_entreprise,omitempty"`
 }
 
 // Sirene detail
@@ -508,15 +399,6 @@ func (e etablissement) intoBatch(batch *pgx.Batch) {
 	}
 }
 
-func contains(array []string, test string) bool {
-	for _, s := range array {
-		if s == test {
-			return true
-		}
-	}
-	return false
-}
-
 func (s scoreFile) toLibelle() string {
 	months := map[string]string{
 		"01": "Janvier",
@@ -547,27 +429,27 @@ func (s scoreFile) toLibelle() string {
 func importHandler(c *gin.Context) {
 	tx, err := db.Get().Begin(context.Background())
 	if err != nil {
-		c.AbortWithError(500, err)
+		c.AbortWithError(http.StatusInternalServerError, err)
 		return
 	}
 	log.Print("preparing import (truncate tables)")
 	err = prepareImport(&tx)
 	if err != nil {
-		c.AbortWithError(500, err)
+		c.AbortWithError(http.StatusInternalServerError, err)
 		return
 	}
 	sourceEntreprise := viper.GetString("sourceEntreprise")
 	log.Printf("processing entreprise file %s", sourceEntreprise)
 	err = processEntreprise(sourceEntreprise, &tx)
 	if err != nil {
-		c.AbortWithError(500, err)
+		c.AbortWithError(http.StatusInternalServerError, err)
 		return
 	}
 	sourceEtablissement := viper.GetString("sourceEtablissement")
 	log.Printf("processing etablissement file %s", sourceEtablissement)
 	err = processEtablissement(sourceEtablissement, &tx)
 	if err != nil {
-		c.AbortWithError(500, err)
+		c.AbortWithError(http.StatusInternalServerError, err)
 		return
 	}
 	// log.Print("refreshing materialized views")
@@ -581,7 +463,7 @@ func importHandler(c *gin.Context) {
 	log.Print("drop dead data")
 	_, err = db.Get().Exec(context.Background(), "vacuum;")
 	if err != nil {
-		c.AbortWithError(500, err)
+		c.AbortWithError(http.StatusInternalServerError, err)
 		return
 	}
 }
