+++ src/utils/utils.go
@@ -1,0 +1,10 @@
+package utils
+
+func Contains(array []string, test string) bool {
+	for _, s := range array {
+		if s == test {
+			return true
+		}
+	}
+	return false
+}
